// src/app/atendimento-paciente/ficha-consulta-retorno/shared/model/avaliacao-consumo-habitual.model.ts
export class AvaliacaoConsumoHabitual {

    id: number;
    energiaKcal: string;
	proteinaTotalGramas: string;
	proteinaAVBGramas: string;
	carboidratoGramas: string;
	lipideoTotalGramas: string;
	lipideoSaturadoGramas: string;
	lipideoPoliinsaturadoGramas: string;
	lipideoMonoinsaturadoGramas: string;
	colesterolMiligramas: string;
	fibraGramas: string;
	calcioGramas: string;
	fosforoGramas: string;
	ferroMiligramas: string;
	zincoMiligramas: string;
	magnesioMiligramas: string;
	potassioMiligramas: string;
	vitaminaB6Miligramas: string;
	vitaminaCMiligramas: string;
	leucinaMiligramas: string;
}